+++ Engine/Source/Renderer/RHI/Resources/RHITexture.h
@@ -31,8 +31,8 @@ class RHITexture2D
 {
 public:
 	int32_t NrChannels = 0;
-	int32_t Width = 0;
-	int32_t Height = 0;
+	uint64_t Width = 0;
+	uint64_t Height = 0;
 	uint16_t NumMips = 0;
 
 	// TODO: Debug only
